[v1.3.32] refactor legacy components and update some components
