+++ src/pages/positions/components/RegisterPositionForm.tsx
@@ -4,7 +4,7 @@ import { useGetSiteResponsiblesMutation } from "../../../services/userService";
 import { useEffect, useState } from "react";
 import { Responsible } from "../../../data/user/user";
 import Strings from "../../../utils/localizations/Strings";
-import UserSelectionModal from "../../../components/UserSelectionModal";
+import UserSelectionModal from "../../components/UserSelectionModal";
 import { UserOutlined } from "@ant-design/icons";
 
 interface FormProps {
