+++ src/pages/site/SitesPage.tsx
@@ -16,7 +16,7 @@ import {
 import { UnauthorizedRoute } from "../../utils/Routes";
 import MainContainer from "../../pagesRedesign/layout/MainContainer";
 import useCurrentUser from "../../utils/hooks/useCurrentUser";
-import PaginatedList from "../../components/PaginatedList";
+import PaginatedList from "../components/PaginatedList";
 import SiteForm, { SiteFormType } from "./components/SiteForm";
 import SiteCard from "./components/SiteCard";
