+++ src/pages/site/components/SiteFormCard.tsx
@@ -29,7 +29,7 @@ import {
   BsTelephone,
   BsTelephoneOutbound,
 } from "react-icons/bs";
-import AnatomyTooltip from "../../../components/AnatomyTooltip";
+import AnatomyTooltip from "../../components/AnatomyTooltip";
 import Strings from "../../../utils/localizations/Strings";
 import {
   FIREBASE_IMAGE_FILE_TYPE,
